# Conflicts:
#	src/com/sys/data/book/BookformProductData.java
#	src/com/sys/entity/_hibernate.cfg.xml

+++ src/com/sys/dao/ScoreDao.java
@@ -18,7 +18,7 @@ public class ScoreDao extends HibernateDao<Score> {
 		super(Score.class);
 	}
 
-	public int getAllScore(String sysUserId, Date startDate, Date endDate) {
+	public double getAllScore(String sysUserId, Date startDate, Date endDate) {
 		
 		StringBuffer sql = new StringBuffer(" select coalesce(sum(s.score),0) " +
 				"from t_score s where s.sys_user_id =:sysUserId ");
@@ -40,27 +40,27 @@ public class ScoreDao extends HibernateDao<Score> {
 			query.setParameter("endDate", endDate);
 		}
 		
-		return Integer.valueOf(query.uniqueResult().toString());
+		return Double.valueOf(query.uniqueResult().toString());
 	}
 
-	public int getFactoryHistoryScore(String factoryId) {
+	public double getFactoryHistoryScore(String factoryId) {
 		StringBuffer sql = new StringBuffer(" select coalesce(sum(s.score),0) " +
 			"from t_score s where s.factory_id =:factoryId and score>=0");
 		
 		Query query = this.getSession().createSQLQuery(sql.toString()).
 		setParameter("factoryId", factoryId);
 
-		return Integer.valueOf(query.uniqueResult().toString());
+		return Double.valueOf(query.uniqueResult().toString());
 	}
 	
-	public int getFactoryConsumeScore(String factoryId) {
+	public double getFactoryConsumeScore(String factoryId) {
 		StringBuffer sql = new StringBuffer(" select coalesce(sum(s.score),0) " +
 			"from t_score s where s.factory_id =:factoryId and score <0");
 		
 		Query query = this.getSession().createSQLQuery(sql.toString()).
 		setParameter("factoryId", factoryId);
 
-		return Integer.valueOf(query.uniqueResult().toString());
+		return Double.valueOf(query.uniqueResult().toString());
 	}
 	
 }
